Reading directly to git repository.
My previous line of work kind of imploded.  The cleanups that were happening
were really geared to having a one pass svndumpfile output.  Some of these
may come back in a different form.  MergeParentData was breaking the output
for me by attempting to do too much.

This code has lots of ad-hoc debugging junk in it.  I'm commiting it
because it seems to do the filesystem part OK.  I need to do RESTOREs now
which means that I've got to actually start commiting files into the repo,
which means I need to work more on scheduling.