Filter out ADDs for moved projects.
ADD records seem to be duplicated in the parent project of the
destination of a moved project. Not good for us, since that directory
will already be created when we move causing any subsequent modifications
to that directory to fail because the files will be in the wrong directory
(e.g. path/to/foo (correct) vs. path/to/foo/foo (incorrect)).
So remove the child that ADDs.